requirements.txt
# Core
aiogram==3.4.1
yt-dlp==2024.12.23
python-dotenv==1.0.0

# Database
aiosqlite>=0.19.0

# Async utilities
aiofiles>=23.0.0

# Rate limiting
aiolimiter>=1.1.0

# Logging
python-json-logger>=2.0.0
